backend/requirements.txt
fastapi==0.116.1
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.2
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
scipy==1.16.1
sympy==1.13.1
tqdm==4.66.5
uvicorn==0.35.0
gunicorn==23.0.0